% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI.R
\name{BMI_function}
\alias{BMI_function}
\title{test}
\usage{
BMI_function(weight, height)
}
\arguments{
\item{weight}{numeric value of waist circumference in cm}

\item{height}{numeric value of hip circumference in cm}
}
\value{
Body Mass index
}
\description{
test
}
\examples{
BMI <- BMI_function(80, 1.7)
}
