% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WHR.R
\name{WHR_function}
\alias{WHR_function}
\title{test}
\usage{
WHR_function(waist_lenght, hip_lenght)
}
\arguments{
\item{waist_lenght}{numeric value of waist circumference in cm}

\item{hip_length}{numeric value of hip circumference in cm}
}
\value{
Waist hip ratio
}
\description{
test
}
\examples{
WHR <- WHR_function(90, 95)
}
